requirements.txt
-i https://download.pytorch.org/whl/cu117
--extra-index-url https://pypi.org/simple

tqdm == 4.58.0
requests == 2.25.1
importlib-metadata == 3.7.0
filelock == 3.0.12
sklearn == 0.0
tokenizers == 0.10.1
explainaboard_client == 0.0.7
pyyaml == 6.0.1
wandb == 0.15.8
pandas == 2.0.3
optuna == 3.3.0

torch == 2.0.0+cu117
torchvision == 0.15.1+cu117
torchaudio == 2.0.1